modify ValidateLogs to be a function in a common package
